requirements.txt
matplotlib==3.6.1
pandas==1.5.3
Pillow==9.4.0
scikit_learn==1.2.1
torch==1.13.0
torchvision==0.14.1
tqdm==4.64.0
